🔧 refactor(xandr_android): make XandrAdListener class open for extension 🎨 style(xandr.g.kt): improve code readability by adjusting indentation and line breaks ✨ feat(xandr.g.kt): add loadInterstitialAd and showInterstitialAd methods to XandrHostApi interface 🔧 refactor(XandrPlugin.kt): implement loadInterstitialAd and showInterstitialAd methods ✨ feat(xandr_android.dart): add loadInterstitialAd and showInterstitialAd methods to XandrAndroid class 🔧 refactor(messages.dart): add loadInterstitialAd and showInterstitialAd methods to XandrHostApi abstract class
The changes were made to add support for interstitial ads in the Xandr Android package. The XandrAdListener class was made open for extension to allow creating a custom listener for interstitial ads. The loadInterstitialAd and showInterstitialAd methods were added to the XandrHostApi interface, implemented in the XandrPlugin.kt file, and exposed in the XandrAndroid class. The same methods were also added to the XandrHostApi abstract class in messages.dart. Code readability was improved in xandr.g.kt by adjusting indentation and line breaks.