🔧 refactor(xandr): add lazy loading support for banner ads
🔧 refactor(xandr): refactor ad listener classes for better code organization

🔧 refactor(xandr): update BannerViewOptions to include enableLazyLoad option

🔧 refactor(xandr): update BannerAd class to support lazy loading

🔧 refactor(xandr): update XandrAdListener to support lazy loading

🔧 refactor(xandr): update XandrInterstitialAdListener to support lazy loading

🔧 refactor(xandr): update XandrBannerAdListener to support lazy loading

🔧 refactor(xandr): update main.dart to include enableLazyLoad option

These changes were made to improve the performance of the application by enabling lazy loading for banner ads. This defers the creation of the internal ad view until the ad is actually loaded, reducing the initial load time of the application. The ad listener classes were also refactored for better code organization.